added title to signup, login. refactor signup
